Replace <=> comparison with is_a? for SNMP::EndOfMibView check
Within the <=> comparable method of the SNMP::Integer class, the
‘other’ side of the comparison is hit with to_i.  In these three cases
when checking to see if `value == SNMP::EndOfMibView`,
`EndOfMibView.to_i` is called and raises an error which is no longer
handled with Ruby 2.3.

I have simply changed the three instances of `value ==
SNMP::EndOfMibView` to use `value.is_a? SNMP::EndOfMibView` instead,
which is a valid comparison regardless of the returned object.

I don’t know the underlying reason for checking for this class to begin
with, so this may not be the appropriate solution.